Rename package from deployment to codegen
