+++ extensions/grpc/codegen/src/main/resources/META-INF/services/io.quarkus.deployment.CodeGenProvider
@@ -1,1 +1,1 @@
-io.quarkus.grpc.deployment.GrpcCodeGen
+io.quarkus.grpc.codegen.GrpcCodeGen
